safari-600.7-branch/LayoutTests/media/video-paused-0-rate-expected.txt


Test that setting '0' playbackRate works correctly.

EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.paused == 'false') OK
END OF TEST
